feat: add H3 block
