+++ src/components/BlockMenu.vue
@@ -121,6 +121,11 @@ const options = computed(() => {
       callback: () => setBlockType(BlockType.H2),
     }, {
       type: 'Turn into',
+      icon: 'bi-type-h3',
+      label: 'Heading 3',
+      callback: () => setBlockType(BlockType.H3),
+    }, {
+      type: 'Turn into',
       icon: 'bi-hr',
       label: 'Divider',
       callback: () => setBlockType(BlockType.Divider),
