+++ src/index.css
@@ -2,3 +2,9 @@
 @tailwind base;
 @tailwind components;
 @tailwind utilities;
+
+[contenteditable]:empty:after {
+  content: attr(data-ph);
+  color: rgba(0, 0, 0, 0.2);
+  cursor: text;
+}
